Add files via upload
These datasets aim to analyze variables that are likely to influence the business across verticals like Banking and Restaurant Delivery using basic python concepts of for-loops and user-defined functions along with data manipulation and visualization techniques using Pandas and NumPy libraries. The goal is to prepare the learner to understand the basic ways to analyze business datasets and draw preliminary conclusions that would pave the way for more extensive Exploratory Data Analysis.